refactor(undo): split checkpoint side effects
